+++ cmd/vacuum_report.go
@@ -4,26 +4,30 @@
 package cmd
 
 import (
+	"bytes"
+	"compress/gzip"
 	"encoding/json"
 	"errors"
+	"fmt"
 	"github.com/daveshanley/vacuum/cui"
 	"github.com/daveshanley/vacuum/model"
 	"github.com/daveshanley/vacuum/motor"
 	"github.com/daveshanley/vacuum/rulesets"
 	"github.com/pterm/pterm"
 	"github.com/spf13/cobra"
+	"gopkg.in/yaml.v3"
 	"io/ioutil"
 	"os"
 	"time"
 )
 
-func GetSpectralReportCommand() *cobra.Command {
+func GetVacuumReportCommand() *cobra.Command {
 
-	return &cobra.Command{
-		Use:   "spectral-report",
-		Short: "Generate a Spectral compatible JSON report",
-		Long: "Generate a JSON report using the same model as Spectral. Default output " +
-			"filename is 'vacuum-spectral-report.json' located in the working directory.",
+	cmd := &cobra.Command{
+		Use:   "report",
+		Short: "Generate a vacuum report",
+		Long: "Generate a full report of a linting run. This can be used as a result set, or can be used to replay a linting run. " +
+			"the default filename is 'vacuum-report-MM-DD-YY-HH_MM_SS.json' located in the working directory.",
 		Example: "vacuum report my-awesome-spec.yaml <vacuum-spectral-report.json>",
 		RunE: func(cmd *cobra.Command, args []string) error {
 
@@ -36,8 +40,17 @@ func GetSpectralReportCommand() *cobra.Command {
 			}
 
 			timeFlag, _ := cmd.Flags().GetBool("time")
+			useYaml, _ := cmd.Flags().GetBool("yaml")
+			noPretty, _ := cmd.Flags().GetBool("no-pretty")
+			compress, _ := cmd.Flags().GetBool("compress")
+			rulesetFlag, _ := cmd.Flags().GetString("ruleset")
 
-			reportOutput := "vacuum-spectral-report.json"
+			extension := ".json"
+
+			if useYaml {
+				extension = ".yaml"
+			}
+			reportOutput := "vacuum-report"
 
 			if len(args) > 1 {
 				reportOutput = args[1]
@@ -53,8 +66,6 @@ func GetSpectralReportCommand() *cobra.Command {
 				pterm.Println()
 				return fileError
 			}
-
-			rulesetFlag, _ := cmd.Flags().GetString("ruleset")
 
 			// read spec and parse to dashboard.
 			defaultRuleSets := rulesets.BuildDefaultRuleSets()
@@ -92,26 +103,67 @@ func GetSpectralReportCommand() *cobra.Command {
 
 			duration := time.Since(start)
 
-			// serialize
-			spectralReport := resultSet.GenerateSpectralReport(args[0]) // todo: convert to full path.
+			// pre-render
+			resultSet.PrepareForSerialization(ruleset.SpecInfo)
 
-			data, err := json.MarshalIndent(spectralReport, "", "    ")
+			var data []byte
+			var err error
+
+			// create vacuum report
+			vr := model.VacuumReport{
+				Generated: time.Now(),
+				SpecInfo:  ruleset.SpecInfo,
+				ResultSet: resultSet,
+			}
+
+			if !useYaml {
+				if noPretty || compress {
+					data, err = json.Marshal(vr)
+				} else {
+					data, err = json.MarshalIndent(vr, "", "    ")
+				}
+			} else {
+				data, err = yaml.Marshal(vr)
+			}
 
 			if err != nil {
-				pterm.Error.Printf("Unable to read marshal report into JSON '%s': %s\n", args[0], fileError.Error())
+				pterm.Error.Printf("Unable to marshal report '%s': %s\n", args[0], fileError.Error())
 				pterm.Println()
 				return err
 			}
 
-			err = ioutil.WriteFile(reportOutput, data, 0664)
+			reportData := data
+
+			if compress {
+
+				var b bytes.Buffer
+				gz := gzip.NewWriter(&b)
+				if _, err = gz.Write(data); err != nil {
+					pterm.Error.Printf("Unable to compress report '%s': %s\n", args[0], err.Error())
+					pterm.Println()
+					return err
+				}
+				if err = gz.Close(); err != nil {
+					pterm.Error.Printf("Unable to close compressed report '%s': %s\n", args[0], err.Error())
+					pterm.Println()
+					return err
+				}
+				reportData = b.Bytes()
+				extension = ".gz"
+			}
+
+			reportOutputName := fmt.Sprintf("%s-%s%s",
+				reportOutput, vr.Generated.Format("01-02-06-15_04_05"), extension)
+
+			err = ioutil.WriteFile(reportOutputName, reportData, 0664)
 
 			if err != nil {
-				pterm.Error.Printf("Unable to write report file: '%s': %s\n", reportOutput, fileError.Error())
+				pterm.Error.Printf("Unable to write report file: '%s': %s\n", reportOutputName, fileError.Error())
 				pterm.Println()
 				return err
 			}
 
-			pterm.Info.Printf("Report generated for '%s', written to '%s'\n", args[0], reportOutput)
+			pterm.Info.Printf("Report generated for '%s', written to '%s'\n", args[0], reportOutputName)
 			pterm.Println()
 
 			fi, _ := os.Stat(args[0])
@@ -120,5 +172,8 @@ func GetSpectralReportCommand() *cobra.Command {
 			return nil
 		},
 	}
-
+	cmd.Flags().BoolP("compress", "c", false, "Compress results using gzip")
+	cmd.Flags().BoolP("yaml", "y", false, "Render using YAML instead of JSON")
+	cmd.Flags().BoolP("no-pretty", "n", false, "Render JSON with no indenting (not available for YAML)")
+	return cmd
 }
